Only unique shares
In order for the Shamir's Secret to work in every case is it necessary to distribute only unique shares. That means each share should have a unique x coordinate.

Function check_duplicates deleted. Numbers are checked before putting them in the x_coordinates set.

Added exception case: if the number of total shares is bigger than the maximum number of shares, we throw an error. (In order, for the shares to be unique we have to make sure that we can add another distinct share)