Upload and download Dropbox functionalities
SSSS_DropBox.py provides the upload/download functionality of shares to Dropbox. A few auxiliary functions have been added.

content_upload.txt is holds for a few seconds the content (shares) that has to be uploaded to the cloud Dropbox .

dropbox_config.py is the file in which the authentication to Dropbox takes place.

API.py is the entry point of the application.