Changed design
Major changes:
The application will begin with a new start page where the dealer can decide if he wants to distribute a secret or reconstruct a secret.

The distribution of  the secret "path" will ask the dealer to input the secret, number of shares and threshold in the input stage, then he will be moved to the setup of cloud databases stage where he will have to input how many shares will be distributed to each of the cloud services chosen at the same stage.

If for example, 3 participants want to use Firebase and 1 CleverCloud, they can do so.